[Modified] OC-1076 #comment updated with 4.5.1 branch #time 1h

+++ src/frontend/src/entities/connection/components/components/general/change_component/form_elements/form_connection/form_methods/mapping/enhancement/Enhancement.js
@@ -96,7 +96,7 @@ class Enhancement extends Component {
 	 * to update expert code
 	 */
 	updateExpertCode(code, e) {
-		if (code.length <= Validation.TextLength.Short) {
+		if (code.length <= Validation.TextLength.Long) {
 			const { setEnhancement } = this.props;
 			let { enhancement } = this.props;
 			enhancement.expertCode = code;
